backend/api-backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
pydantic>=2.7.0
email-validator>=2.1.0
pydantic-settings>=2.2.1
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
APScheduler>=3.10.4
httpx>=0.27.0
python-dotenv>=1.0.1
openai>=1.20.0
gunicorn>=21.2.0
sentry-sdk>=2.0.0
redis>=5.0.0
